backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.6.0
PyMySQL==1.1.0
python-dotenv==1.0.1
